test(explorer): add tests for transfers table
