--- apps/explorer/src/app/routes/treasury/network-treasury.spec.tsx
@@ -1,5 +1,0 @@
-describe('NetworkTreasury', () => {
-  it('renders headers correctly', () => {
-    expect(true).toBe(true);
-  });
-});
