+++ apps/explorer/tsconfig.app.json
@@ -10,7 +10,9 @@
   },
   "files": [
     "../../node_modules/@nx/react/typings/cssmodule.d.ts",
-    "../../node_modules/@nx/react/typings/image.d.ts"
+    "../../node_modules/@nx/react/typings/image.d.ts",
+    "src/app/routes/treasury/components/network-accounts-table.tsx",
+    "src/app/routes/treasury/components/network-transfers-table.tsx"
   ],
   "exclude": [
     "**/*.spec.ts",
